Add the `watchProposal()` function in PFOffer
Adding the watchProposal() function that @alexvandesande introduced in
his Sample Offer contract
[here](https://gist.github.com/alexvandesande/122c05a0f8a7108523aed29ae7ec92b1#file-proposal-js-L271).

This way there is no requirement to pass the `proposalID` as a
constructor argument. The contractor should instead add it on his own
after he makes the proposal by calling `watchProposal()`.

If the contractor tries to cheat by making 2 or more different proposals and
adding the one that has gotten the most votes as the proposal ID in
`watchProposal()` the token holders will notice and raise the alert so
that the contractor can be blacklisted.